+++ app/src/auth.lua
@@ -13,6 +13,10 @@ local function auth(app)
 	app:match("post_comment", "/post/:post_id[%d]/comment", r2(require("actions.comment_create")))
 	app:match("create_subreddit", "/subreddit/create", r2(require("actions.create_subreddit")))
 	app:match("subscribe", "/subscribe/:subreddit", r2(require("actions.subscribe")))
+	app:match("comment_edit", "/comment/:comment_id[%d]/edit", r2(require("actions.comment_edit")))
+	app:match("comment_delete", "/comment/:comment_id[%d]/delete", r2(require("actions.comment_delete")))
+	app:match("post_edit", "/post/:post_id[%d]/edit", r2(require("actions.post_edit")))
+	app:match("post_delete", "/post/:post_id[%d]/delete", r2(require("actions.post_delete")))
 
 	-- app:match("prefs", "/prefs", function(self) end) -- stub